requirements.txt
python-pptx>=0.6.22
Pillow>=10.0.0
